Add handling for attaching/detaching USB devices
During runtime it is now possible to attach/detach a USB device.
The RecyclerView on the main page will adjust accordingly by
adding or removing the appropriate elements.
The adapter for the RecyclerView now holds the data in an ArrayList
instead of an Array in order to perform add and remove actions.
Reinitialize the timer every time (on startReceive) because it
can't be resschulded once canceled.
Back button in toolbar on Operations.java didn't call onBackPressed(),
fixed that.

Note: something is still not working  when the user selects a device,
then goes back to the main screen (with the device still plugged in)
and selects that same device again. RequestWait() fails under these
circumstances.