Add the basic construct for a MidiHandlerService
This was done because of the need to sustain the communication with the
USB-device, even when the app isn't in the foreground or when the screen
is turned off.
Therefore we introduce a MidiHandlerService responsible for exactly
this job. It is a service that pushes itself into the foreground, so
that it is less likely to be killed by the Android system. Foreground
services need to show a permanent notification, indiciating to the user
that something is happening in the background.
Note that this state is just the basic construct, it doesn't include any
logic to the MidiHandlerService. This is going to be done in the future.