Improve usb-input data query and processing
Move the 'heavy' processing in the MidiHandlerService from the main
thread to another thread by using a ScheduldedExecutorService instead
of a Timer/TimerTask. This also gives the ability to specify nanoseconds
instead of microseconds for the period.
Improve the processing itself by adjusting the matcher for the received
data (from the USB-device). It can happend that the keyboard sends
multiple commands in one bulk transfer, therefore we just search for the
last appearance of the specified regex-pattern in the data.