Improve thread handling
The MyUsbDeviceConnection now throws a CalledFromWrongThreadException,
if send(...) or receive() is called from the UI thread. This prevents
that the programmer accidentally blocks the UI; from now on we need to
call these methods on a worker thread.
The new behaviour was then implemented in MidiHandlerService and
Operations, e. g. try-catch and starting the new thread.