Introduce spinners to update recv/send messages
The user can now select predefined MidiMessages by clicking on different
items in two spinners. To fill the spinners with data, a whole package
was created, named 'midimessage'. This package includes classes needed
to specify different MidiMessages (for the Tyros). The
MidiControllerBuilder constructs these messages (by using the other
classes) and returns them as a list. To update the recv/send messages,
the spinner asks for the MidiController, which gives access to the
MidiMessage and therefore the hex message to recv/send.
The MidiMessages also include a property specifiying alternative
MidiChannels. They will be used in the future to let the user pick (in
the spinner items) from different MidiChannels, e. g. if they want to
send a 'Volume'-message for RIGHT_1, RIGHT_2, RIGHT_3 or/and left.