Improve OperationRules to allow for multiple rules
Therefore an OperationRulesManager was created which stores the
OperationRules. The MidiHandlerService now processes the data for each
OperationRule separately and successively.
At the moment, regarding the spinners, there should only be one
OperationRule. When the user selects another item in one of the
spinners, a new OperationRule is created and the old one is deleted.
(Outlook: maybe we should change this behaviour, so that an
OperationRule can be changed later on.)