Customize spinners with 'settings'-icon
Create a CustomMidiControllerAdapter to show a 'settings'-icon next to
the text in the spinners, but only when they are collapsed. When the
user wants to select another MidiController, we don't need this
'settings'-option.
(Outlook: At the moment, when the user clicks on the 'settings'-icon,
we just display a toast. In the future, the app should show a Dialog to
select the MidiChannels the spinner should work with.)