Init dialog for custom MidiChannel-option in spinner
Create a dialog when the user clicks on the settings icon in the
spinner. In this dialog, the user can select which channels to use
for the OperationRules. The available options (MidiChannels) are passed
to the dialog via an object tag in the ImageButton (gear-icon for
settings). To fill the dialog with readable options to select from,
add a readable name to the MidiChannels in MidiConstants.
Note that until now, there is no logic for when the user selects
something in the dialog. Next steps are to adapt the OperationRules to
our requirements.