Implement custom MidiChannels-logic ('settings'-icon)
Add the logic to the 'settings'-icon. When the user clicks on it, an
AlertDialog opens up where the user can select the MidiChannels to work
with.
To be more flexible, the meaning of OperationRule changed! Beforehand,
an OperationRule described 'Volume', 'Panpot' etc. in the spinner. Now
an OperationRule is the whole thing, including a spinner to select the
receive MidiChannelControllers and the send MidiChannelControllers.
Therefore, the OperationRule doesn't save String messages anymore, but
MidiChannelMessages, in order to be able to remove these messages later
on as well.
Outlook: This way, we can simply add a new OperationRule later. We will
do this by adding a RecyclerView to the whole layout. At the moment, we
just initialize a new OperationRule in Operations onCreate() method.
Due to these changes, the logic of Midi processing in
MidiHandlerService also had to change accordingly. Furthermore, when
the user clicks on 'Update' or 'OK' in the dialog, we need to update
the list in the dialog and the OperationRule. This logic was
implemented in Operations.
Note that we changed the MidiController to MidiChannelController, in
order to use its method without casting all the time. We will later
seek for a more generic solution to also implement other
MidiChannelMessages (at the moment: only Control Change
MidiChannelMessages are supported).