Init RecyclerView for OperationRules
Add a RecyclerView to Operations-layout. The RecyclerView has a
RulesAdapter including a data set of RulesViewModels. Every
RulesViewModel is associated with one item in the RecyclerView, meaning
one OperationRule. When we init the RecyclerView in Operations, we fill
it with a basic RulesViewModel, so that there is already one element in
the list.
Via the RulesViewModel the RulesViewHolder can bin data to the spinners
by setting the appropriate SpinnerAdapters. Furthermore, we specified
the layout including the two spinners the RulesViewHolder holds
referencese to.
We also outsourced the SpinnerListeners from Operations to a new class
'SpinnerListener', responsible for handling the onItemSelected-event
(user selected an item in the spinner) and the onClick-event (user
pressed the 'settings'-icon in a spinner).
From now on, empty OperationRules can only be initialized via the
OperationRulesManager. This allows us to always add a newly
constructed OperationRule to the overall list of OperationRules, which
is crucial for the request from MidiHandlerService.
(Outlook: We haven't implemented a add functionality yet, although we
have set up the methods to change the data set in the RecyclerView.
Add a button to Operations later.)