Implement add/delete feature for OperationRules-RecyclerView
The user can now click on a FloatingActionButton (FAB) in order to add
a new OperationRule to the list. To implement this FAB, add the
appropriate dependency to the Gradle build file in the app-folder.
Furthermore, add an ImageButton to delete an OperationRule from the
list. There are two listeners to listen to events from this list:
 - RulesListener is responsible for handling the event when the user
   clicks on the delete button to delete an OperationRule.
 - SpinnerListener is responsible for handling the event when the user
   clicks on another item in the Spinner, thus changing the current
   OperationRule and the event when the user clicks on the 'settings'-
   button to specify the MidiChannels for the current OperationRule.